Revert adding date_genotyped to Litter
After pushing that commit to github, I started getting a server error on rmb-colony-staging every time I tried to go anywhere past the main admin page. I suspect this is because there are now unmigrated changes to the database. I was going to try making the migrations locally, pushing to github then running migrate on Heroku, but I got an error when trying to make migrations locally; I need some additional module. In the meantime, I'm reverting this change to check whether I can get rid of the server error.